Change the button placement and appearance, and fix some bugs.
The first change is the button size. I noticed that the left control panel had a lot of unused space, so I increased the button size to better fill the area.

The second change is a fix for the slider bug. In the original version, the slider would send incorrect values to the playground and kuksa-client if the user moved it too quickly. This issue has been fixed in this update.

The final change is related to the position value. In this update, the position value displayed on dk_ivi will not exceed 3.

Signed-off-by: dan2-ux <danhlanhat9@gmail.com>